requirements.txt
berserk
python-chess
stockfish